backend/requirements.txt
gunicorn==20.1.0
requests==2.26.0
Django==3.2.3
djangorestframework==3.12.4
django-filter==23.1
djoser==2.1.0
Pillow==9.0.0
psycopg2-binary==2.9.3
